Fix embedded aliases for direct plugin loading
